add build OK check to deploy script

+++ deploy.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
-docker build -t kntu.py .
-docker run -it --rm --name kntu.py kntu.py
 
+# build docker image
+docker build -t kntu.py .
+BUILD_OK=$?
+
+# if we built OK, run it
+if [ $BUILD_OK -eq 0 ]; then
+    docker run -it --rm --name kntu.py kntu.py
+fi
